Add pings clj-session (brain-dump).
It parses my ~/.pings file which contains minute-wise
`date --rfc-3339=seconds` lines. It's been quite interesting to see what
I've been up to.

However, to be actually useful the data would have to be combined with
some "was working" state, such as #key-events, name of the application
running or a combination of this.

This is definitively inspired by [Steven Wolfram's Quantified Self
post](http://blog.stephenwolfram.com/2012/03/the-personal-analytics-of-my-life/).