shopping_list: What, where and how much.
This version has the basic grouping (category, initial, price, weight)
in place but the frontend, the sorting and a storage backend are
lacking.

Dedicated to my little sister, thus intended to eventually demonstrate
at least the concepts to here and an exercise in creating beautiful
abstractions in a functional language.

At least one criterion is partly fulfilled: it's a simple solution and
so far it works well. It's almost like an algebra which is quite
pleasing.

Well now, have fun, take care an' good night.