chunking demo using three.js.
far slower than trixl itself, not sure why yet. even with frustum
culling disabled, it seems that the communication with GL is still too
much. possibly because we use a custom material for each cube?

still, if we could batch-draw each chunk i think this should be far
faster as we'd only draw 64 batches for a -100,100 area.