structs (and tuple structs, and also newtypes)
a tuple struct is basically a named tuple.  as a struct it has a name
and as a tuple it doesn't have field names.

tuple structs with only one field are also called newtypes (or said to
be using the newtype pattern).  they are useful for wrapping values to
be able to differentiate them from other values with the same field
type.