Fix spawning of the error reporting goroutine
It has to be spawned *before* we run the main loop, because otherwise if
we had too many errors this would have blocked them.  (Where too many is
`2 * config.concurrency`.  After *all* fetching goroutines would be
blocked and could never resume because the `errors` channel is buffered
as well.  That shouldn't happen now.)