Get much closer to get something real up and running
This is about a day (and a half?) of work, mostly spent trying to
understand what planck is doing, and then porting code over,
occasionally chasing segfaults.

We can evaluate ClojureScript expressions now, but not require the
planck namespaces yet.  The errors point to missing files, but I'm
guessing that some of the `function_*` ports are incomplete.