Write to the same directory as the input file
This only happens if the output file is specified in the YAML prefix, no
output file was given on the commandline, the specified output file is
only a file name *and* the input file is not in the current directory.

It ensures that `blog /some/path/to/blog.yaml` doesn't generate a file
in the current directory if the output was specified as `output: file.yaml`.