Fix serving PDFs
They were converted to text and served in a `<pre>` element, because I
did not consider non-text things.  Now they are just served fresh from
the disk.

Maybe we should switch the logic around so that non-html is the default,
but... not yet. :)

(I usually don't write code that I don't need yet, so... that code isn't
here yet.)