Fix truncated command output
The issue was that poll_fds contained state: When one of the fds was
closed then it would be set to -1, which usually would make `poll()`
ignore it.  But because we recreated poll_fds with every call to
`poll()` we would not save that state, then try to remove it again and
increment `dead_fds` twice for one fd which meant it would not check the
other (still possibly open) fd.

🎉