grounds: Fix drawing text in grid view example with system font
`drawTextInRect` will **not**draw the text at all if the text does not
fit in the box!!!  This is of course font dependent, but it stumped me
for quite a while.

The example included in the docs does not work out of the box, but the
example that comes as a file with the SDK _does_ work because there
they use a custom font.

Whoops, that was a bit of a trip.