waves: Parse _something_ via WebAssembly from HTML/JS :)
Main problem (took me several hours) was to figure out why the Uint8Array
passed in was empty for the longest time when it arrived in Zig land.

Problem is that you can only pass various number types to WebAssembly
and thus need to work with pointers, essentially.  (That second article
below may be worth reading again.)

So you need to have a `WebAssembly.Memory`, coordinate that with your
compiler of whatever to wasm and then use it correctly.  A bunch of
steps!

Next: Try to figure out how to get some structs passed two and fro'.
Probably also via pointers?  Bit of a mystery.  Maybe that's what
those binding generators are for. 🤔

Helpful resources:

- https://rob-blackbourn.github.io/blog/2020/06/07/wasm-arrays/
- https://blog.mjgrzymek.com/blog/zigwasm